// src/polls/store/types.ts
export enum PollsMutationEnum {
  SET_COOPERATION_POLLS = 'SET_COOPERATION_POLLS',
  SET_SELECTED_POLL = 'SET_SELECTED_POLL',
  ADD_COOPERATION_POLL = 'ADD_COOPERATION_POLL',
  DELETE_POLL = 'DELETE_POLL',
  UPDATE_POLL = 'UPDATE_POLL',
  SET_POll_BY_ID = 'SET_POll_BY_ID',
  SET_COOPERATION_POLLS_WITH_RESULTS = 'SET_COOPERATION_POLLS_WITH_RESULTS',
}

export enum PollsActionEnum {
  SET_COOPERATION_POLLS = 'SET_COOPERATION_POLLS',
  SET_POll_BY_ID = 'SET_POll_BY_ID',
  SET_SELECTED_POLL = 'SET_SELECTED_POLL',
  ADD_COOPERATION_POLL = 'ADD_COOPERATION_POLL',
  GET_POll_BY_ID = 'GET_POll_BY_ID',
  DELETE_POLL = 'DELETE_POLL',
  UPDATE_POLL = 'UPDATE_POLL',
  SET_COOPERATION_POLLS_WITH_RESULTS = 'SET_COOPERATION_POLLS_WITH_RESULTS',
}

export enum PollsGettersEnum {
  getPolls = 'getPolls',
  getSelectedPoll = 'getSelectedPoll',
  getPollByID = 'getPollByID',
  getPollsWithResults = 'getPollsWithResults',
}

export interface PollsStateInterface {
  cooperationPolls: Array<PollModel>;
  cooperationPollsWithResults: Array<PollModelWithResults>;
  selectedPoll: PollModel;
  displayModal: boolean;
}
// ...
export type Mutations<S = PollsStateInterface> = {
  [PollsMutationEnum.SET_COOPERATION_POLLS](state: S, payload: Array<PollModel>): void;
  [PollsMutationEnum.SET_COOPERATION_POLLS_WITH_RESULTS](state: S, payload: Array<PollModelWithResults>): void;
  [PollsMutationEnum.SET_SELECTED_POLL](state: S, payload: number): void;
  [PollsMutationEnum.ADD_COOPERATION_POLL](state: S, payload: PollModel): void;
  [PollsMutationEnum.DELETE_POLL](state: S, payload: number): void;
  [PollsMutationEnum.UPDATE_POLL](state: S, payload: EditPollPayloadInterface): void;
  [PollsMutationEnum.SET_POll_BY_ID](state: S, payload: PollModel): void;
};

export interface Actions {
  [PollsActionEnum.SET_COOPERATION_POLLS]({ commit }: AugmentedActionContext, payload: number): void;
  [PollsActionEnum.SET_COOPERATION_POLLS_WITH_RESULTS]({ commit }: AugmentedActionContext, payload: number): void;
  [PollsActionEnum.SET_SELECTED_POLL]({ commit }: AugmentedActionContext, payload: number): void;
  [PollsActionEnum.ADD_COOPERATION_POLL]({ commit }: AugmentedActionContext, payload: any): void;
}
// ...
export type AugmentedActionContext = {
  commit<K extends keyof Mutations>(key: K, payload: Parameters<Mutations[K]>[1]): ReturnType<Mutations[K]>;
} & Omit<ActionContext<PollsStateInterface, RootStateInterface>, 'commit'>;